refactor: unify handling of payment submissions
