+++ src/payment/_style.scss
@@ -1,4 +1,4 @@
-@import "./apple-pay/apple-pay";
+@import "./payment-methods/apple-pay/apple-pay";
 
 
 @keyframes skeleton-pulse {
