--- src/payment/apple-pay/index.jsx
@@ -1,6 +1,0 @@
-export { default as ApplePayButtonContainer } from './ApplePayButtonContainer';
-export {
-  configureApiService,
-  performApplePayPayment,
-  redirectToReceipt,
-} from './service';
