--- src/payment/cybersource/index.jsx
@@ -1,4 +1,0 @@
-export { default as reducer } from './data/reducers';
-export { default as saga } from './data/sagas';
-export { submitPaymentCybersource } from './data/actions';
-export { configureApiService } from './data/service';
