+++ src/payment/data/sagas.js
@@ -1,4 +1,4 @@
-import { all, call, put, takeEvery } from 'redux-saga/effects';
+import { call, put, takeEvery, select } from 'redux-saga/effects';
 
 // Actions
 import {
@@ -7,14 +7,22 @@ import {
   addCoupon,
   removeCoupon,
   updateQuantity,
+  submitPayment,
 } from './actions';
 
 // Services
 import * as PaymentApiService from './service';
 
-import { saga as cybersourceSaga } from '../cybersource';
-import { saga as payPalSaga } from '../paypal';
+import { checkout as checkoutCybersource } from '../payment-methods/cybersource';
+import { checkout as checkoutPaypal } from '../payment-methods/paypal';
+import { checkout as checkoutApplePay } from '../payment-methods/apple-pay';
 import { handleErrors, handleMessages } from '../../feedback';
+
+export const paymentMethods = {
+  cybersource: checkoutCybersource,
+  paypal: checkoutPaypal,
+  'apple-pay': checkoutApplePay,
+};
 
 export function* handleFetchBasket() {
   try {
@@ -88,14 +96,39 @@ export function* handleUpdateQuantity({ payload }) {
   }
 }
 
+export function* handleSubmitPayment({ payload }) {
+  const { method, ...paymentArgs } = payload;
+  try {
+    const paymentMethodCheckout = paymentMethods[method];
+    const basket = yield select(state => ({ ...state.payment.basket }));
+    yield put(submitPayment.request());
+    const result = yield call(paymentMethodCheckout, basket, paymentArgs);
+    yield put(submitPayment.success(result));
+  } catch (error) {
+    if (error.aborted === true) {
+      // This an a user aborted action, do nothing
+    } else {
+      yield put(submitPayment.failure(error));
+      if (error.code) {
+        // Client side generated errors are simple error objects and
+        // here we wrap them into the same format as the api returns them in
+        yield call(handleErrors, { errors: [error] }, true);
+      } else {
+        yield call(handleErrors, error, true);
+      }
+      if (error.basket) {
+        yield put(basketDataReceived(error.basket));
+      }
+    }
+  } finally {
+    yield put(submitPayment.fulfill());
+  }
+}
+
 export default function* saga() {
   yield takeEvery(fetchBasket.TRIGGER, handleFetchBasket);
   yield takeEvery(addCoupon.TRIGGER, handleAddCoupon);
   yield takeEvery(removeCoupon.TRIGGER, handleRemoveCoupon);
   yield takeEvery(updateQuantity.TRIGGER, handleUpdateQuantity);
-
-  yield all([
-    cybersourceSaga(),
-    payPalSaga(),
-  ]);
+  yield takeEvery(submitPayment.TRIGGER, handleSubmitPayment);
 }
