+++ src/payment/payment-methods/apple-pay/ApplePay.messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from '@edx/frontend-i18n';
+
+const messages = defineMessages({
+  'payment.apple.pay.pay.with.apple.pay': {
+    id: 'payment.apple.pay.pay.with.apple.pay',
+    defaultMessage: 'Pay with Apple Pay',
+    description: 'The title of an apple pay button',
+  },
+});
+
+export default messages;
