+++ src/payment/payment-methods/apple-pay/_apple-pay.scss
