+++ src/payment/payment-methods/apple-pay/index.jsx
@@ -1,0 +1,6 @@
+export { default as ApplePayButton } from './ApplePayButton';
+export {
+  configureApiService,
+  checkout,
+  redirectToReceipt,
+} from './service';
