+++ src/payment/payment-methods/paypal/assets/paypal-logo.png
