+++ src/payment/payment-methods/paypal/index.jsx
@@ -1,0 +1,2 @@
+export { default as PayPalButton } from './PayPalButton';
+export { configureApiService, checkout } from './service';
