--- src/payment/paypal/index.jsx
@@ -1,4 +1,0 @@
-export { default as PayPalButton } from './PayPalButton';
-export { default as reducer } from './data/reducers';
-export { default as saga } from './data/sagas';
-export { configureApiService } from './data/service';
